+++ src/plugin/ui/about_window.cpp
@@ -1,7 +1,7 @@
 //---------------------------------------------------------------------------------------------------------------------
 //   XMidiCtrl - MIDI Controller plugin for X-Plane
 //
-//   Copyright (c) 2021-2022 Marco Auer
+//   Copyright (c) 2021-2023 Marco Auer
 //
 //   XMidiCtrl is free software: you can redistribute it and/or modify it under the terms of the
 //   GNU Affero General Public License as published by the Free Software Foundation, either version 3
@@ -54,7 +54,7 @@ void about_window::create_widgets()
     ImGui::NewLine();
     ImGui::Separator();
     ImGui::NewLine();
-    ImGui::TextUnformatted("Copyright (c) 2021-2022 by Marco Auer");
+    ImGui::TextUnformatted("Copyright (c) 2021-2023 by Marco Auer");
     ImGui::NewLine();
     ImGui::TextUnformatted(XMIDICTRL_NAME " has been released under the GNU Affero General Public License");
     ImGui::TextUnformatted("in hope that it will be useful, but without any warranty.");